+++ src/main.rs
@@ -155,13 +155,11 @@ async fn main() -> Result<()> {
         }
         Commands::Checkout {
             target,
-            history,
-            since,
             combined_view,
             output,
         } => {
             let ctx = commands::Context::new(cli.verbose, &target, cfg).await?;
-            commands::checkout::run(&ctx, history, since, combined_view, &output).await
+            commands::checkout::run(&ctx, combined_view, &output).await
         }
         Commands::Sync {
             target,